# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Dianping Agent Platform API
  version: 1.3.0
servers:
  - url: http://localhost:8081
security:
  - bearerAuth: []
paths:
...
  /api/v1/customer-service/imports/preview:
    post:
      summary: Upload workbook and create a dry-run import preview
      description: Requires cs:data:import. Parses into staging only; no business facts are written before confirm.
      parameters:
        - { $ref: '#/components/parameters/TenantHeader' }
        - { $ref: '#/components/parameters/WorkspaceHeader' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Preview result with counts, warnings and confirmability
          content: { application/json: { schema: { $ref: '#/components/schemas/CsImportBatch' } } }
        '422':
          description: CS_IMPORT_VALIDATION_FAILED
          content: { application/json: { schema: { $ref: '#/components/schemas/AiError' } } }
  /api/v1/customer-service/imports/{batchId}:
    parameters:
      - { $ref: '#/components/parameters/BatchIdPath' }
      - { $ref: '#/components/parameters/TenantHeader' }
      - { $ref: '#/components/parameters/WorkspaceHeader' }
    get:
      summary: Get import batch status and counts
      responses:
        '200':
          description: Batch state
          content: { application/json: { schema: { $ref: '#/components/schemas/CsImportBatch' } } }
  /api/v1/customer-service/imports/{batchId}/errors:
    parameters:
      - { $ref: '#/components/parameters/BatchIdPath' }
      - { $ref: '#/components/parameters/TenantHeader' }
      - { $ref: '#/components/parameters/WorkspaceHeader' }
    get:
      summary: Page masked validation errors of an import batch
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Error page (values are masked; no raw PII)
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PageResponse' }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/CsImportError' }
  /api/v1/customer-service/imports/{batchId}/confirm:
    parameters:
      - { $ref: '#/components/parameters/BatchIdPath' }
      - { $ref: '#/components/parameters/TenantHeader' }
      - { $ref: '#/components/parameters/WorkspaceHeader' }
    post:
      summary: Confirm a previewed batch for staged commit preparation
      description: Requires cs:data:import. DATA-003 only moves an eligible batch to CONFIRMING; DATA-004 performs the idempotent fact commit.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CsImportConfirmRequest' }
      responses:
        '200':
          description: Commit preparation accepted; created/updated/skipped remain zero until DATA-004
          content: { application/json: { schema: { $ref: '#/components/schemas/CsImportCommitResult' } } }
        '409':
          description: CS_IMPORT_CONFLICT (hash/parserVersion/status/TTL mismatch)
          content: { application/json: { schema: { $ref: '#/components/schemas/AiError' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    TenantHeader: { name: X-Tenant-Id, in: header, required: true, schema: { type: string, maxLength: 64 } }
    WorkspaceHeader: { name: X-Workspace-Id, in: header, required: true, schema: { type: string, maxLength: 64 } }
...
    BatchIdPath: { name: batchId, in: path, required: true, schema: { type: string, maxLength: 64 } }
...
  schemas:
    AiError:
      type: object
      properties:
        success: { type: boolean }
        code:
          oneOf:
            - { type: string }
            - { type: integer }
        errorMsg: { type: string }
        message: { type: string }
    PageResponse:
      type: object
      required: [items, total, page, size]
      properties:
        items: { type: array, items: { type: object } }
        total: { type: integer, format: int64, minimum: 0 }
        page: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1 }
...
    CsImportBatchStatus:
      type: string
      enum: [PREVIEWING, READY_TO_CONFIRM, REJECTED, CONFIRMING, CONFIRMED, CANCELLED, EXPIRED]

    # --- Customer service payloads --------------------------------------------
    CsImportCounts:
      type: object
      additionalProperties: false
      required: [consumerAliases, conversations, messages, orderSnapshots, serviceCases, sourceLinks, missingMedia]
      properties:
        consumerAliases: { type: integer, minimum: 0 }
        conversations: { type: integer, minimum: 0 }
        messages: { type: integer, minimum: 0 }
        orderSnapshots: { type: integer, minimum: 0 }
        serviceCases: { type: integer, minimum: 0 }
        sourceLinks: { type: integer, minimum: 0 }
        missingMedia: { type: integer, minimum: 0 }
    CsImportBatch:
      type: object
      required: [batchId, status, sourceSha256, parserVersion, counts, warningCount, errorCount, blockingErrorCount, confirmable, expiresAt, version]
      properties:
        batchId: { type: string }
        status: { $ref: '#/components/schemas/CsImportBatchStatus' }
        fileName: { type: string }
        sourceSha256: { type: string, minLength: 64, maxLength: 64 }
        parserVersion: { type: string }
        counts: { $ref: '#/components/schemas/CsImportCounts' }
        warningCount: { type: integer, minimum: 0 }
        errorCount: { type: integer, minimum: 0, description: Blocking errors; kept alongside blockingErrorCount for the DATA-003 operator contract }
        blockingErrorCount: { type: integer, minimum: 0 }
        confirmable: { type: boolean }
        expiresAt: { type: string, format: date-time }
        version: { type: integer, minimum: 0 }
        warnings:
          type: array
          maxItems: 50
          items: { type: string, maxLength: 500 }
        createdAt: { type: string, format: date-time }
        confirmedAt: { type: string, format: date-time, nullable: true }
    CsImportError:
      type: object
      required: [sheet, row, errorCode, severity, message]
      properties:
        sheet: { type: string }
        row: { type: integer, minimum: 0 }
        field: { type: string, nullable: true }
        errorCode: { type: string }
        severity: { type: string, enum: [WARNING, BLOCKING] }
        maskedValue: { type: string, nullable: true, description: Sensitive values are always masked }
        message: { type: string }
    CsImportConfirmRequest:
      type: object
      additionalProperties: false
      required: [expectedSourceSha256, expectedParserVersion, expectedVersion]
      properties:
        expectedSourceSha256: { type: string, minLength: 64, maxLength: 64 }
        expectedParserVersion: { type: string }
        expectedVersion: { type: integer, minimum: 0 }
        warningsReviewed: { type: boolean, default: false, description: Must be true when the batch has warnings }
    CsImportCommitResult:
      type: object
      required: [batchId, status, version, created, updated, skipped]
      properties:
        batchId: { type: string }
        status: { $ref: '#/components/schemas/CsImportBatchStatus' }
        version: { type: integer, minimum: 0 }
        created: { $ref: '#/components/schemas/CsImportCounts' }
        updated: { $ref: '#/components/schemas/CsImportCounts' }
        skipped: { $ref: '#/components/schemas/CsImportCounts' }
...